Move ech decoding to another crate
