+++ Cargo.toml
@@ -24,6 +24,7 @@ doctest = false
 members = [
   "dns",
   "dns-transport",
+  "ech-config",
 ]
