+++ ech-config/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "ech-config"
+version = "0.1.0"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+# logging
+log = "0.4"
+
+# protocol parsing helper
+byteorder = "1.3"
+
+# printing of certain packets
+base64 = "0.13.0"
+pretty_assertions = "0.7.2"
+
+[dev-dependencies]
+pretty_assertions = "0.7"
+env_logger = "0.9.0"
+
